add classic structure

+++ uninstall.php
@@ -1,8 +1,10 @@
 <?php
+// If uninstall is not called, exit
+if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+    exit;
+}
 
-defined( 'WP_UNINSTALL_PLUGIN' ) || exit;
-
-$deletable_options = [ 'thrail-wp_activated', 'thrail-wp_db_version' ];
+$deletable_options = [ 'thrail-wp_version', 'thrail-wp_install_time', 'thrail-wp_docs_json', 'codexpert-blog-json' ];
 foreach ( $deletable_options as $option ) {
     delete_option( $option );
 }
